Factor out some common things into shared helpers (examples crate)
