+++ examples/src/bin/wallet.rs
@@ -1,48 +1,29 @@
 use std::error::Error;
-use std::{env, time, vec};
-use tkhq_api_key_stamper::TurnkeyApiKey;
+use std::{env, vec};
 use tkhq_client::generated::immutable::common::v1::{HashFunction, PayloadEncoding};
-use tkhq_client::generated::{CreateWalletIntent, DeleteWalletsIntent, SignRawPayloadIntentV2};
 use tkhq_client::generated::{
     immutable::common::v1::{AddressFormat, Curve, PathFormat},
     WalletAccountParams,
 };
+use tkhq_client::generated::{CreateWalletIntent, DeleteWalletsIntent, SignRawPayloadIntentV2};
+use tkhq_examples::{current_time_ms, load_api_key_from_env};
 
 // See <https://docs.turnkey.com/api-reference/organizations/create-sub-organization> for documentation
 const TURNKEY_API_HOST: &str = "https://api.turnkey.com";
 
 #[tokio::main]
 async fn main() -> Result<(), Box<dyn Error>> {
-    // Load .env file from the example folder (examples/.env)
-    let current_dir = env::current_dir()?; // should be the workspace root
-    let env_path = current_dir.join("examples").join(".env");
+    // Load API key
+    let api_key = load_api_key_from_env()?;
 
-    if env_path.exists() {
-        dotenvy::from_path(&env_path)?;
-    } else {
-        println!("No .env file found at {:?}", env_path);
-        println!("Continuing because env might already be populated with the right variables");
-    }
-
-    let api_public_key =
-        env::var("TURNKEY_API_PUBLIC_KEY").expect("cannot load TURNKEY_API_PUBLIC_KEY");
-    let api_private_key =
-        env::var("TURNKEY_API_PRIVATE_KEY").expect("cannot load TURNKEY_API_PRIVATE_KEY");
-
-    let api_key = TurnkeyApiKey {
-        private_key_hex: api_private_key,
-        public_key_hex: api_public_key.clone(),
-    };
-
-    // Create the request body for our create_wallet request
+    // Get organization ID from env
     let organization_id =
         env::var("TURNKEY_ORGANIZATION_ID").expect("cannot load TURNKEY_ORGANIZATION_ID");
-    let timestamp_ms = time::SystemTime::now()
-        .duration_since(time::UNIX_EPOCH)
-        .unwrap()
-        .as_millis();
 
+    // Create our Turnkey client
     let client = tkhq_client::TurnkeyClient::new(TURNKEY_API_HOST, api_key, None);
+
+    // Create the activity intent
     let intent = CreateWalletIntent {
         wallet_name: "New wallet".to_string(),
         accounts: vec![WalletAccountParams {
@@ -55,7 +36,7 @@ async fn main() -> Result<(), Box<dyn Error>> {
     };
 
     let res = client
-        .create_wallet(organization_id.clone(), timestamp_ms, intent)
+        .create_wallet(organization_id.clone(), current_time_ms(), intent)
         .await?;
 
     assert_eq!(res.addresses.len(), 1);
@@ -71,7 +52,7 @@ async fn main() -> Result<(), Box<dyn Error>> {
     let signature_activity_result = client
         .sign_raw_payload(
             organization_id.clone(),
-            timestamp_ms,
+            current_time_ms(),
             SignRawPayloadIntentV2 {
                 sign_with: eth_address.to_string(),
                 payload: "hello from TKHQ".to_string(),
@@ -92,7 +73,7 @@ async fn main() -> Result<(), Box<dyn Error>> {
     let delete_wallet_result = client
         .delete_wallets(
             organization_id,
-            timestamp_ms,
+            current_time_ms(),
             DeleteWalletsIntent {
                 wallet_ids: vec![wallet_id.clone()],
                 delete_without_export: Some(true),
